(long) Double.parseDouble

+++ src/filesprocessing/filters/BetweenFilter.java
@@ -6,8 +6,8 @@ public class BetweenFilter extends Filter {
 	private static final int SIZE_FACTOR = 1024;
 
 	public boolean passFilter(File f, String[] args) {
-		long upper_bound = Long.parseLong(args[1].replace(" ",""));
-		long lower_bound = Long.parseLong(args[2].replace(" ",""));
+		long upper_bound = (long) Double.parseDouble(args[1].replace(" ",""));
+		long lower_bound = (long) Double.parseDouble(args[2].replace(" ",""));
 		long size = f.length() / SIZE_FACTOR;
 
 		return (lower_bound >= 0 && upper_bound >= lower_bound && lower_bound <= size && size <= upper_bound);
