+++ src/filesprocessing/filters/SmallerThanFilter.java
@@ -5,7 +5,7 @@ import java.io.File;
 public class SmallerThanFilter extends Filter{
 	private static final int SIZE_FACTOR = 1024;
 	public boolean passFilter(File f, String[] args){
-		long bound =  Long.parseLong(args[1].replace(" ",""));
+		long bound =  (long) Double.parseDouble(args[1].replace(" ",""));
 		long size = f.length()/SIZE_FACTOR;
 		return (bound>=0 && size < bound);
 	}
